Move to ui-utils existing func
