+++ compose/ui/ui-util/src/wasmJsMain/kotlin/androidx/compose/ui/util/Trace.wasmJs.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 The Android Open Source Project
+ * Copyright 2024 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,6 +16,6 @@
 
 package androidx.compose.ui.util
 
-//actual inline fun <T> trace(sectionName: String, block: () -> T): T {
-//    return block()
-//}
+actual inline fun <T> trace(sectionName: String, block: () -> T): T {
+    return block()
+}
